2-Ollama Chatbot/app.py
import streamlit as st
from langchain_community.llms import Ollama
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser
import os

from dotenv import load_dotenv
load_dotenv()

## Langsmith Tracking

os.environ["LANGCHAIN_API_KEY"] = os.getenv("LANGCHAIN_API_KEY")
os.environ["LANGCHAIN_TRACKING"] = 'true'
os.environ["LANGCHAIN_PROJECT"] = "Ollama QnA Chatbot"

## Prompt Template

prompt = ChatPromptTemplate.from_messages(
    [
        ("system", "You are a helpful assistant. Please give resonse to the user query."),
        ("user", "Question: {question}")
    ]
)

def generate_response(question, engine, temperature, max_tokens):
    llm = Ollama(model=engine)
    output_parser = StrOutputParser()
    chain = prompt|llm|output_parser
    answer = chain.invoke({'question': question})
    return answer


## Title of the app

st.title("Ollama QnA Chatbot")

## Sidebar Settings
engine = st.sidebar.selectbox("Select an Ollama Model", ["llama3.2:latest", "gemma3:latest"])

## Adjust Response Parameter
temperature = st.sidebar.slider("Temperature", min_value=0.0, max_value=1.0, value=0.7)
max_tokens = st.sidebar.slider("Max Tokens", min_value= 50, max_value = 300, value= 150)

## Main interface for user
st.write("Ask Anything")
user_input = st.text_input("You:")

if user_input:
    response = generate_response(user_input, engine, temperature, max_tokens)
    st.write(response)

else:
    st.write("Please provide your query")
