requirements.txt
langchain_openai
langchain
langchain_community
python-dotenv
langchain_core
streamlit
openai
ollama
langchain-groq
ipykernel
pypdf
pymupdf
wikipedia
langchain-text-splitters
chromadb
faiss-cpu
langchain_chroma
duckdb
pandas
sentence_transformers
langchain_huggingface
Arxiv
Wikipedia
tqdm
huggingface_hub[hf_xet]
duckduckgo_search
ddgs
SQLAlchemy
validators==0.28.1
youtube-transcript-api
unstructured
pytube
numexpr
huggingface_hub
gradio